# Improve line, col calculate performance by use move cursor on Pairs Iterator.