--- a/grammars/benches/json.rs
+++ b/grammars/benches/json.rs
@@ -56,16 +56,37 @@ item = _{ SOI ~ line* ~ EOI }
     pub struct JsonParser;
 }
 
+// With 500 times iter
+// pair.line_col                                time:   [2.9937 µs 2.9975 µs 3.0018 µs]
+// position.line_col                            time:   [212.59 µs 213.38 µs 214.29 µs]
+// position.line_col (with fast-line-col)       time:   [18.241 µs 18.382 µs 18.655 µs]
+//
+// With 1000 times iter
+// pair.line_col                                time:   [10.814 µs 10.846 µs 10.893 µs]
+// position.line_col                            time:   [90.135 µs 93.901 µs 98.655 µs]
+// position.line_col (with fast-line-col)       time:   [1.7199 ms 1.7246 ms 1.7315 ms]
 fn line_col_benchmark(c: &mut Criterion) {
     let mut file = File::open("benches/main.i18n.json").unwrap();
     let mut data = String::new();
 
     file.read_to_string(&mut data).unwrap();
     let pairs = autocorrect::JsonParser::parse(autocorrect::Rule::item, &data).unwrap();
-    let last_pair = pairs.last().unwrap();
-    c.bench_function("line col", |b| {
+
+    c.bench_function("pair.line_col", |b| {
         b.iter(|| {
-            last_pair.line_col();
+            let mut pairs = pairs.clone();
+            for _ in 0..500 {
+                pairs.next().unwrap().line_col();
+            }
+        })
+    });
+
+    c.bench_function("position.line_col", |b| {
+        b.iter(|| {
+            let mut pairs = pairs.clone();
+            for _ in 0..500 {
+                pairs.next().unwrap().as_span().start_pos().line_col();
+            }
         });
     });
 }