--- a/pest/src/iterators/flat_pairs.rs
+++ b/pest/src/iterators/flat_pairs.rs
@@ -12,7 +12,6 @@ use alloc::vec::Vec;
 use core::fmt;
 
 use super::pair::{self, Pair};
-use super::pairs::{Cursor, CursorPairs};
 use super::queueable_token::QueueableToken;
 use super::tokens::{self, Tokens};
 use crate::RuleType;
@@ -29,7 +28,6 @@ pub struct FlatPairs<'i, R> {
     input: &'i str,
     start: usize,
     end: usize,
-    cursor: Cursor,
 }
 
 /// # Safety
@@ -46,7 +44,6 @@ pub unsafe fn new<R: RuleType>(
         input,
         start,
         end,
-        cursor: Cursor::default(),
     }
 }
 
@@ -110,16 +107,7 @@ impl<'i, R: RuleType> Iterator for FlatPairs<'i, R> {
             return None;
         }
 
-        let pair = unsafe {
-            pair::new(
-                Rc::clone(&self.queue),
-                self.input,
-                self.start,
-                self.cursor.clone(),
-            )
-        };
-        self.move_cursor(pair.as_str());
-
+        let pair = unsafe { pair::new(Rc::clone(&self.queue), self.input, self.start, None) };
         self.next_start();
 
         Some(pair)
@@ -134,14 +122,7 @@ impl<'i, R: RuleType> DoubleEndedIterator for FlatPairs<'i, R> {
 
         self.next_start_from_end();
 
-        let pair = unsafe {
-            pair::new(
-                Rc::clone(&self.queue),
-                self.input,
-                self.end,
-                self.cursor.clone(),
-            )
-        };
+        let pair = unsafe { pair::new(Rc::clone(&self.queue), self.input, self.end, None) };
 
         Some(pair)
     }
@@ -162,18 +143,7 @@ impl<'i, R: Clone> Clone for FlatPairs<'i, R> {
             input: self.input,
             start: self.start,
             end: self.end,
-            cursor: self.cursor.clone(),
         }
-    }
-}
-
-impl<'i, R: RuleType> CursorPairs for FlatPairs<'i, R> {
-    fn cursor(&self) -> Cursor {
-        self.cursor.clone()
-    }
-
-    fn cursor_mut(&mut self) -> &mut Cursor {
-        &mut self.cursor
     }
 }
 
@@ -205,5 +175,25 @@ mod tests {
                 .collect::<Vec<Rule>>(),
             vec![Rule::c, Rule::b, Rule::a]
         );
+    }
+
+    #[test]
+    fn test_line_col() {
+        let mut pairs = AbcParser::parse(Rule::a, "abcNe\nabcde").unwrap().flatten();
+
+        let pair = pairs.next().unwrap();
+        assert_eq!(pair.as_str(), "abc");
+        assert_eq!(pair.line_col(), (1, 1));
+        assert_eq!(pair.line_col(), pair.as_span().start_pos().line_col());
+
+        let pair = pairs.next().unwrap();
+        assert_eq!(pair.as_str(), "b");
+        assert_eq!(pair.line_col(), (1, 2));
+        assert_eq!(pair.line_col(), pair.as_span().start_pos().line_col());
+
+        let pair = pairs.next().unwrap();
+        assert_eq!(pair.as_str(), "e");
+        assert_eq!(pair.line_col(), (1, 5));
+        assert_eq!(pair.line_col(), pair.as_span().start_pos().line_col());
     }
 }