--- a/pest/src/iterators/pair.rs
+++ b/pest/src/iterators/pair.rs
@@ -20,7 +20,7 @@ use core::str;
 #[cfg(feature = "pretty-print")]
 use serde::ser::SerializeStruct;
 
-use super::pairs::{self, Pairs};
+use super::pairs::{self, Cursor, Pairs};
 use super::queueable_token::QueueableToken;
 use super::tokens::{self, Tokens};
 use crate::span::{self, Span};
@@ -43,6 +43,9 @@ pub struct Pair<'i, R> {
     input: &'i str,
     /// Token index into `queue`.
     start: usize,
+
+    line: usize,
+    col: usize,
 }
 
 /// # Safety
@@ -52,11 +55,14 @@ pub unsafe fn new<R: RuleType>(
     queue: Rc<Vec<QueueableToken<R>>>,
     input: &str,
     start: usize,
+    cursor: Cursor,
 ) -> Pair<'_, R> {
     Pair {
         queue,
         input,
         start,
+        line: cursor.line,
+        col: cursor.col,
     }
 }
 
@@ -239,6 +245,11 @@ impl<'i, R: RuleType> Pair<'i, R> {
     #[cfg(feature = "pretty-print")]
     pub fn to_json(&self) -> String {
         ::serde_json::to_string_pretty(self).expect("Failed to pretty-print Pair to json.")
+    }
+
+    /// Returns the `line`, `col` of this pair start.
+    pub fn line_col(&self) -> (usize, usize) {
+        (self.line, self.col)
     }
 
     fn pair(&self) -> usize {