ref: #707, #560

This implementation is based by I use in AutoCorrect: https://github.com/huacnlee/autocorrect/blob/v2.5.5/autocorrect/src/result.rs#L267

## Add `line_col` method for get `(line, col)` in pair

```rust
let mut pairs = JSONParser::parse(Rule::a, "....").unwrap();
let pair = pairs.next().unwrap();
let (line, col) = pair.line_col()
```

In this method, line, col has managed by move_cursor on Pairs iterator `next` method. For a large input string this method will be faster, because it does not need to re-calculate all of the input string. The line, col value has been cached in the pair and accumulate by `next` method.

And I keep the old `line_col` method in `Position` struct, because I found that has been used in other without `Pairs` iterator. This method still works, but it will be slower than the new method.

## Benchmark

```
pair.line_col                              time:   [2.9937 µs 2.9975 µs 3.0018 µs]
position.line_col                          time:   [212.59 µs 213.38 µs 214.29 µs]
position.line_col (with fast-line-col)     time:   [18.241 µs 18.382 µs 18.655 µs]
```

> 🎊 9x fast than position.line_col  with `fast-line-col`
> 🌈 100x fast than old position.line_col  

## Limitation

This implementation only work for forward iteration for `Pairs`. So iterate backward will fallback to `Position::line_col` method.