@huacnlee 2022-12-29
We still need `fast-line-col`, because my implementation only work for call from Pairs.

There have many case use `Position` in directly.